Version 9.0.4.
Chris sent me this tarball today, but it dates from 2012.  The CHANGELOG is
short ("Fixed yaap measure (Stephen Robertson)") but there seem to be a
lot of changes in a lot of files from 9.0a3, including several measure files.
This commit just brings in his code without even compile checking or running tests.

+++ get_qrels.c
@@ -122,8 +122,8 @@ te_get_qrels (EPI *epi, char *text_qrels_file, ALL_REL_INFO *all_rel_info)
     while (*ptr) {
 	if (UNDEF == parse_qrels_line (&ptr, &line_ptr->qid,
 				       &line_ptr->docno, &line_ptr->rel)) {
-	    fprintf (stderr, "trec_eval.get_qrels: Malformed line %d\n",
-		     line_ptr - lines + 1);
+	    fprintf (stderr, "trec_eval.get_qrels: Malformed line %ld\n",
+		     (long) (line_ptr - lines + 1));
 	    return (UNDEF);
 	}
 	line_ptr++;
